requirements.txt
pillow
pytesseract
googletrans==3.1.0a0
tk